% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mikeplot-package.R
\docType{package}
\name{mikeplot-package}
\alias{mikeplot}
\alias{mikeplot-package}
\title{mikeplot: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mikemahoney218/mikeplot}
  \item \url{https://mikemahoney218.github.io/mikeplot/}
  \item Report bugs at \url{https://github.com/mikemahoney218/mikeplot/issues}
}

}
\author{
\strong{Maintainer}: Michael Mahoney \email{mike.mahoney.218@gmail.com} (\href{https://orcid.org/0000-0003-2402-304X}{ORCID})

}
\keyword{internal}
