% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{annotation_north_arrow}
\alias{annotation_scale}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ggspatial}{\code{\link[ggspatial]{annotation_north_arrow}}, \code{\link[ggspatial]{annotation_scale}}}
}}

