% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_pub}
\alias{theme_pub}
\alias{theme_map}
\title{Opinionated ggplot2 themes}
\usage{
theme_pub(base_fill = "white", base_size = 11, base_family = "")

theme_map(
  base_line_color = "grey80",
  base_fill = "white",
  base_size = 11,
  base_family = ""
)
}
\arguments{
\item{base_fill}{The color to use for both plot and panel background.}

\item{base_size}{The base text size, used for axis labels. All other elements
are sized in proportion to this value.}

\item{base_family}{The font family to use.}
}
\value{
An object of classes \code{theme} and \code{gg}.
}
\description{
Opinionated ggplot2 themes
}
\examples{
ggplot(mtcars, aes(wt, mpg, color = factor(cyl))) +
  geom_point() +
  labs(
    title = "Fuel economy declines as weight increases",
    subtitle = "Larger engines are generally heavier"
   ) +
  theme_pub()

}
